Most current version of Team yellow's  Love you A Latte
Has a working log and retrieve time button form the DB, as well a three item list menu on every page as the 1st div.  At the time of this upload we have completed 19 stories points.  JS for the product page needs to be added to start to implement PHP or JS shopping cart.